purple-haze-python/requirements.txt
colour-science>=0.4.4
matplotlib>=3.8
numpy>=1.24
